requirements.txt
#django
django
djangorestframework
django-environ
drf-yasg

flake8
validate_email_address

#pytest
pytest
pytest-django
pytest-cov
